refactored ama and created ama-module
