+++ src/app/ama/ask-me-anything/ama-event/ama-event.component.css
