+++ src/app/ama/ask-me-anything/ask-me-anything-page/ask-me-anything-page.component.ts
@@ -1,5 +1,5 @@
 import { MessageBoxComponent } from '../message-box/message-box.component';
-import { ChatMessage } from 'src/app/static/chat-message/chat-message';
+import { ChatMessage } from 'src/app/models/chat-message/chat-message';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
 import { environment } from 'src/environments/environment';
 import * as Stomp from 'stompjs';
