+++ src/app/ama/ask-me-anything/message-box/message-box.component.css
